update startup code
