--- src/Client.Infrastructure/Extensions/ApplicationBuilderExtensions.cs
@@ -1,8 +1,0 @@
-﻿namespace FSH.BlazorWebAssembly.Client.Infrastructure.Extensions;
-public static class ApplicationBuilderExtensions
-{
-    public static IApplicationBuilder UseInfrastructure(this IApplicationBuilder app, IConfiguration config)
-    {
-        return app;
-    }
-}
