+++ src/Client.Infrastructure/Extensions/HubExtensions.cs
@@ -9,12 +9,11 @@ public static class HubExtensions
         if (hubConnection == null)
         {
             hubConnection = new HubConnectionBuilder()
-                              .WithUrl($"{apiBaseUri}notifications", options =>
-                              {
-                                  options.AccessTokenProvider = async () => (await localStorage.GetItemAsync<string>("authToken"));
-                              })
-                              .WithAutomaticReconnect()
-                              .Build();
+                .WithUrl($"{apiBaseUri}notifications", options =>
+                    options.AccessTokenProvider =
+                        async () => await localStorage.GetItemAsync<string>("authToken"))
+                .WithAutomaticReconnect()
+                .Build();
         }
 
         return hubConnection;
